backend/requirements.txt

Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3.1
psycopg2-binary>=2.9
PyZ3950>=2.0
requests>=2.31
gunicorn>=21.2
